Fixed a typo in the MySQL database definition that only shows up when adding resources to courses with IDs higher than 255

+++ mod/resource/db/mysql.sql
@@ -16,7 +16,7 @@
 
 CREATE TABLE prefix_resource (
   id int(10) unsigned NOT NULL auto_increment,
-  course tinyint(10) unsigned NOT NULL default '0',
+  course int(10) unsigned NOT NULL default '0',
   name varchar(255) NOT NULL default '',
   type tinyint(4) NOT NULL default '0',
   reference varchar(255) default NULL,
